Reorganizing Embrapa related classes and functions into a single folder named "embrapa"
