+++ src/main.py
@@ -1,6 +1,6 @@
 from typing import Union, Annotated
-from web_scrapping import get_export_import_page, get_production_commercialization_processing_page
-from classes import ModelExport, ModelImport, ModelProcessing
+from embrapa.web_scrapping import get_export_import_page, get_production_commercialization_processing_page
+from embrapa.enum_models import ModelExport, ModelImport, ModelProcessing
 from fastapi import FastAPI, Path, Depends, HTTPException, status
 from fastapi.security import OAuth2PasswordRequestForm
 from auths.auth import User, get_current_active_user
